Initial contribution of Motorola MOTODEV tools code
This patch contains Eclipse plugins that complement the Android Development
Tools (ADT). The patch also contains the MOTODEV App Validator, a static
analysis tool for Android apps. App Validator is built as an Eclipse RCP
that runs on the command line, but also has an Eclipse IDE integration.  To
build the plugins, you will need Eclipse 3.7 or 4.2 with the PDE installed
with Java 6. An ant build file will come in a separate patch.

Change-Id: I8af2e5980eb24ed1c206bb2ddb9b3d733b9eff56
Signed-off-by: Eric Cloninger <dcp874@motorola.com>