[459363] Enable unit tests.
Works on fixing the unit tests. Adds a git template
for contributors to use when submitting items. Updated
the README with latest items on contributing and working
with the project source code.

Integration tests need an SDK location and items installed
these will continue to fail until we get this setup.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=459363
Signed-off-by: David Carver <d_a_carver@yahoo.com>