[460492] Rename classes and internal strings to reflect changes from previous owners
Deleted the app-validator help as pre-flighting is not part of the project. There were
lots of References to MOTODEV Studio and the com.motorola.studio namespace. These were
changed to use the name Andmore or Eclipse Andmore as appropriate. The namespace became
org.eclipse.andmore.  These changes were in the Dita docs sources as well as Java code
and plugin properties files.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=460492
Signed-off-by: Eric Cloninger <ericc@pobox.com>