[461459] Restrict version of ObjectWeb Dependency
The AOSP lint check tool requires that we use ObjectWeb ASM 4.0.
Any other version will cause NoClassDefFound errors to occur and
the build to lock.   This change makes sure the target platform
has 4.0 available to it, and also makes sure the plugin that
requires this excludes versions below 4.0 and 5.0.  We can't
use higher versions due to breaking API changes.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=461459
Signed-off-by: David Carver <d_a_carver@yahoo.com>