[468473] NPE when expanding application node in db explorer
An NPE would be thrown when trying to show the icon for an application
in the database explorer for the Android Database perspective.
The dependency was missing and the wrong value was hard coded.  Plugin
now uses the correct Constant value instead of a magic value.  So if the
constant value changes this will be picked up automatically.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=468473
Signed-off-by: David Carver <d_a_carver@yahoo.com>