[509663] Add support for aapt --no-version-vectors option.
Setting --no-version-vectors is necessary when using newer versions
(e.g. 25.1.0) of the appcompat library to target older Android devices.
If this flag is not set during the packaging, older devices will not be
able to access Android resource files properly. This results in app
crashes upon startup due to seemingly missing resources.

This commit extends the andmore user interface to include a check box in
the project-specific Android properties page which enables the user to
add the --no-version-vectors option during aapt runs. The value of the
check box is stored in the project.properties file and picked up by the
build helper when running aapt. If the properties file does not contain
the new property, the default is to not add the option during builds in
order to avoid side effects.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=509663
Signed-off-by: Marcus Handte <handte@locoslab.com>