[484566] Fix NPE in BaseBuilder when property file is missing.
When a project does not contain a project.poperties file, the
BaseBuilder will run into a NullPointerException. This change will
prevent the exception and abort the build. 

Note that this does not fix the new project wizard problem that is the
root cause of having an invalid project setup in the first place.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=484566
Signed-off-by: Marcus Handte <handte@locoslab.com>