[509666] Fix NPE in visual editor that causes load failures.
When a project contains locales without region, the visual editor will
fail to load due to a NPE in the algorithm that guesses the best default
locale. As a result, the user will get a "no root view" message.

This change modifies the selection algorithm of the ConfigurationMatcher
to either return a full language and region match or to return the last
matching language setting without causing a NPE when the region is null. 

The only side effect I can see is that it will fully traverse the
list if there is no full match instead of aborting earlier. However, my
feeling is that the resulting "loss of performance" is not noticeable
whereas the NPE clearly is.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=509666
Signed-off-by: Marcus Handte <handte@locoslab.com>